backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
bcrypt>=4.0.1,<5
email-validator==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
pymongo==4.6.0
motor==3.3.2  # Async MongoDB driver

# Redis
redis==5.0.1
hiredis==2.2.3

# Message Queue
celery==5.3.4
kombu==5.3.4

# Kafka
kafka-python==2.0.2
confluent-kafka==2.3.0

# MinIO (S3-compatible storage)
minio==7.2.0

# LLM Providers
openai==1.3.5  # Optional: For OpenAI
groq==0.4.1  # Free, Fast LLM (Llama, Mixtral models)

# ML/AI (Optional - uncomment when implementing ML features)
# Note: These have compatibility issues with Python 3.14+
# For Python 3.14+, use newer versions or Python 3.11/3.12
# scikit-learn==1.4.0  # Newer version supports Python 3.14
# xgboost==2.0.2
# pandas>=2.2.0  # For Python 3.14 support
# numpy>=1.26.0  # For Python 3.14 support

# Email/SMS
sendgrid==6.11.0
twilio==8.10.0

# Template Engine
Jinja2==3.1.2

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
pandas==2.2.0
openpyxl==3.1.2  # For Excel support

# Logging
structlog==23.2.0

# Testing (optional for now)
pytest==7.4.3
pytest-asyncio==0.21.1

